+++ infrastructure/commons/cert_manager/main.tf
@@ -56,3 +56,31 @@ resource "helm_release" "cert_manager_config" {
   depends_on = [helm_release.cert_manager]
 }
 
+#########webhook oci############
+resource "helm_release" "cert_manager_webhook_oci" {
+  count      = var.cloud_provider == "oci" ? 1 : 0
+  name       = "cert-manager-webhook-oci"
+  repository = "https://dn13.gitlab.io/cert-manager-webhook-oci"
+  chart      = "cert-manager-webhook-oci"
+  version    = var.cert_manager_webhook_oci_version
+  namespace  = var.cert_manager_webhook_oci_namespace
+
+  create_namespace  = true
+  disable_webhooks  = false
+  force_update      = true
+  wait              = true
+  wait_for_jobs     = true
+  timeout           = 600
+  atomic            = true
+  cleanup_on_fail   = true
+  replace           = true
+  recreate_pods     = true
+  reset_values      = true
+  reuse_values      = false
+  dependency_update = true
+  max_history       = 10
+
+
+  depends_on = [helm_release.cert_manager]
+}
+